requirements.txt
functions-framework==3.*
aenum==3.1.15
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
debugpy==1.6.7.post1
decorator==5.1.1
Deprecated==1.2.14
exceptiongroup==1.1.3
executing==1.2.0
Flask==2.3.3
frozenlist==1.4.0
future==0.18.3
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2==3.1.2
line-bot-sdk==3.4.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.7
numpy==1.25.2
openai==0.28.0
outcome==1.2.0
packaging==23.1
pandas==2.1.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.10.0
plotly==5.16.1
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.16.1
PyMySQL==1.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.6.2
tenacity==8.2.3
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
Werkzeug==2.3.7
wrapt==1.15.0
wsproto==1.2.0
yarl==1.9.2
zipp==3.16.2
